Fix Mixed-Content errors when using HTTPS
Before this commit, https://scala-lms.github.io/ was broken because
the generated pages contained links to
"https://scala-lms.github.com/..." which are 301-redirected to
"http://scala-lms.github.io/..." and modern browsers refuse to load HTTP
resources in HTTPS pages. The fix is to make sure that the links all
start with "https://scala-lms.github.io/..." so that no redirects-to-HTTP happen.